% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transclustr.R
\name{tclust}
\alias{tclust}
\title{tclust}
\usage{
tclust(dist_obj = NULL, filename = NULL, simmatrix = NULL,
  threshold = NULL, convert_dissimilarity_to_similarity = TRUE,
  file_type = "SIMPLE", normalization = "RELATIVE",
  use_custom_fallback = FALSE, sim_fallback = 0,
  use_default_interval = TRUE, th_min = 0, th_max = 100, th_step = 1,
  p = 1, f_att = 100, f_rep = 100, R = 100, dim = 3, start_t = 100,
  d_init = 0.01, d_maximal = 5, s_init = 0.01, f_s = 0.01,
  fpt_time_limit = 20, fpt_max_cost = 5000, fpt_step_size = 10,
  disable_force = FALSE, disable_fpt = FALSE, seed = 42)
}
\description{
tclust
}
\examples{
library(transclustr)
library(ggplot2)
library(plyr)

SD <- 1.3
CLUSTER_SIZE <- 200
a <- cbind(rnorm(CLUSTER_SIZE, mean = 10, sd = SD), rnorm(n  =CLUSTER_SIZE,  mean = 10,sd = SD))
b <- cbind(rnorm(CLUSTER_SIZE, mean = 0,  sd = SD),  rnorm(n =CLUSTER_SIZE, mean = 10,sd = SD))
c <- cbind(rnorm(CLUSTER_SIZE, mean = 10, sd = SD), rnorm(n  =CLUSTER_SIZE,  mean = 0, sd = SD))
d <- cbind(rnorm(CLUSTER_SIZE, mean = 0,  sd = SD),  rnorm(n =CLUSTER_SIZE, mean = 0, sd = SD))
e <- cbind(rnorm(CLUSTER_SIZE, mean = 5,  sd = SD),  rnorm(n =CLUSTER_SIZE, mean = 5, sd = SD))
#
set <- rbind(a,b,c,d, e)
df_set <- as.data.frame(set)
names(df_set) <- c("x","y")
ggplot(data=df_set,aes(x=x,y=y)) +geom_point()


dist_set <- dist(set)
max(dist_set)

for(norm in c("RELATIVE","ABSOLUTE")){
  tclust_res<- tclust(dist_set, disable_fpt = TRUE, d_maximal = 5,dim = 3,R = 100,threshold = 17,normalization = norm)

  set_res <- as.data.frame(cbind(set,tclust_res$clusters[[1]]))
  names(set_res) <- c("x","y","cluster")

  find_hull <- function(df) df[chull(df$x, df$y), ]
  hulls <- ddply(set_res, "cluster", find_hull)

  length(unique(tclust_res$clusters[[1]]))
  print(ggplot(data=as.data.frame(set_res),aes(x=x,y=y,colour=factor(cluster),fill=factor(cluster))) +
           geom_point() +
           theme(legend.position = "none") +
           geom_polygon(data = hulls,alpha = 0.25) +
           ggtitle(paste("Normalization: ",norm,", Cost: ",tclust_res$costs[[1]]))
  )
}
}

